# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "week 3  API",
    "description": " API for user contact Information and Temple Information",
    "version": "1.0.0",
    "contact": {
      "name": "Developer"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000"
    }
  ],
  "paths": {
...
    "/temple/create": {
      "post": {
        "summary": "Create a new temple",
        "description": "Create a new temple",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "temple_id": {
                    "type": "integer"
                  },
                  "additionalInfo": {
                    "type": "boolean"
                  },
                  "name": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "dedicated": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Temple created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "temple_id": {
                      "type": "integer"
                    },
                    "additionalInfo": {
                      "type": "boolean"
                    },
                    "name": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string"
                    },
                    "dedicated": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/temple/getOne/{id}": {
      "get": {
        "summary": "Get a temple by ID",
        "description": "Get a temple by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the temple to retrieve",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "apiKey",
            "in": "header",
            "required": true,
            "description": "API key for authentication",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Temple retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "_id": {
                      "type": "string"
                    },
                    "temple_id": {
                      "type": "integer"
                    },
                    "additionalInfo": {
                      "type": "boolean"
                    },
                    "name": {
                      "type": "string"
                    },
                    "location": {
                      "type": "string"
                    },
                    "dedicated": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Temple not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/temple/getAll": {
      "get": {
        "summary": "Get all Temples",
        "description": "Get all Temples",
        "parameters": [
          {
            "name": "apiKey",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Temples retrieved successfully",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "_id": {
                    "type": "string"
                  },
                  "temple_id": {
                    "type": "integer"
                  },
                  "additionalInfo": {
                    "type": "boolean"
                  },
                  "name": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "dedicated": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/temple/update/{id}": {
      "put": {
        "summary": "Update a temple with the ID",
        "description": "Update a temple's details by their ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The user ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "temple_id": {
                    "type": "string"
                  },
                  "additionalInfo": {
                    "type": "Boolean"
                  },
                  "name": {
                    "type": "string"
                  },
                  "location": {
                    "type": "string"
                  },
                  "dedicated": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "description": "Success message"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Data to update cannot be empty"
          },
          "404": {
            "description": "User not found"
          },
          "500": {
            "description": "Error updating user"
          }
        }
      }
    },
    "/temple/delete/{id}": {
      "delete": {
        "description": "Delete a temple by ID",
        "summary": "Delete a temple by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Temple deleted successfully"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Temple not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
...
}